[BOD-1052] Refactor
